When a refresh is requested and is already running, return the current refresh.
We maintain a list of ongoing refresh. If a refresh is already
running, just return the deferred attached to it. If we find a more
general refresh, return it too. For example, if we want to refresh a
real server and the corresponding virtual server refresh is running,
return the later instead of doing a partial refresh (which may or may
not be overrided).