[actions] Store actions in database
To avoid to instantiate a SNMP proxy, we store available actions in
database. This information should be current because after each
action, the entity is refreshed.

The SNMP proxy is modified to export a writable attribute which should
be used by plugins to determine if they should populate the action
table.

The plugin interface has been updated. All plugins have been updated.